Fix setup script and the database creation
There were some problems when running the whole setup script
with sudo, since 'source' does not work then and the python
packages were not installed correctly in the virtual environment.

Moreover the creation of the database is now done within the main
program because it can't be part of the setup script since the
Raspberry needs to be rebooted in between for the program to run
without errors.

Also fixed some typos in the README.