Add network module
Possibility to start wsniff in master or slave mode in order to use
multiple RPis as a logical unit.
In order to form a connection, the devices running wsniff have to be in
the same network.
The website/network.py module implements a network discovery using UDP
packets and ensures each party knows the other's IP address after the
connection has been established.
For the master, a new site was added where the User can connect to
waiting devices. It also list all devices you are currently connected
to.